+++ MovieReviewSite.Core/Repositories/Crew/CrewRepository.cs
@@ -1,4 +1,5 @@
 ﻿using Microsoft.EntityFrameworkCore;
+using Microsoft.Identity.Client;
 using Microsoft.IdentityModel.Tokens;
 using MovieReviewSite.Core.Interfaces.ReviewSite;
 using MovieReviewSite.Core.Models;
@@ -15,13 +16,15 @@ public partial class CrewRepository : ICrewRepository
     private readonly ReviewSiteContext _context;
     private readonly IMovieRepository _movieRepository;
     private readonly IReviewRepository _reviewRepository;
+    private readonly IUserRepository _userRepository;
 
     public CrewRepository(ReviewSiteContext context, IMovieRepository movieRepository,
-        IReviewRepository reviewRepository)
+        IReviewRepository reviewRepository, IUserRepository userRepository)
     {
         _context = context;
         _movieRepository = movieRepository;
         _reviewRepository = reviewRepository;
+        _userRepository = userRepository;
     }
 
     public async Task<List<BaseCrew>> GetAllCrew()
@@ -56,43 +59,59 @@ public partial class CrewRepository : ICrewRepository
 
     public async Task AddCrew(NewCrewRequest dto)
     {
-        if (dto.FirstName.IsNullOrEmpty())
+        //only admins or vip members can Add crew so this checks to make sure 
+        var isUserAuthorized = await _userRepository.IsUserAdminOrVIP(dto.CreatedBy);
+        if (isUserAuthorized)
         {
-            throw new ArgumentException("the crew must at least have a firstname!");
+            if (dto.FirstName.IsNullOrEmpty())
+            {
+                throw new ArgumentException("the crew must at least have a firstname!");
+            }
+
+            var crew = new DataBase.Crew()
+            {
+                FirstName = dto.FirstName,
+                LastName = !dto.LastName.IsNullOrEmpty() ? dto.LastName : null,
+                MiddleName = !dto.MiddleName.IsNullOrEmpty() ? dto.MiddleName : null,
+                IsActive = true,
+                IsVisible = true,
+                BirthDate = dto.BirthDate == null ? null : dto.BirthDate,
+                LastModifiedOn = DateTime.UtcNow,
+                CreatedOn = DateTime.UtcNow,
+                CreatedBy = dto.CreatedBy
+            };
+            await _context.Crews.AddAsync(crew);
+            await _context.SaveChangesAsync();
         }
-        var crew = new DataBase.Crew()
-        {
-            FirstName = dto.FirstName,
-            LastName = !dto.LastName.IsNullOrEmpty() ? dto.LastName : null,
-            MiddleName = !dto.MiddleName.IsNullOrEmpty() ? dto.MiddleName : null,
-            IsActive = true,
-            IsVisible = true,
-            BirthDate = dto.BirthDate == null ? null : dto.BirthDate,
-            LastModifiedOn = DateTime.UtcNow,
-            CreatedOn = DateTime.UtcNow,
-            CreatedBy = dto.CreatedBy
-        };
-        await _context.Crews.AddAsync(crew);
-        await _context.SaveChangesAsync();
     }
 
     public async Task UpdateCrew(int id, UpdateCrewRequest dto)
     {
-        var crew = await _context.Crews.Where(c => c.Id == id).SingleOrDefaultAsync();
-        if (crew == null) throw new ArgumentException("crew does not exist!");
-        crew!.FirstName = dto.FirstName.IsNullOrEmpty() ? crew.FirstName : dto.FirstName;
-        crew.LastName = dto.LastName.IsNullOrEmpty() ? crew.LastName : dto.LastName;
-        crew.MiddleName = dto.MiddleName.IsNullOrEmpty() ? crew.MiddleName : dto.MiddleName;
-        crew.BirthDate = dto.BirthDate ?? crew.BirthDate;
-        crew.LastModifiedOn = DateTime.UtcNow;
-        _context.Crews.Update(crew);
-        await _context.SaveChangesAsync();
+        //only admins or vip members can update Crew so this checks to make sure 
+        var isUserAuthorized = await _userRepository.IsUserAdminOrVIP(dto.UserId);
+        if (isUserAuthorized)
+        {
+            var crew = await _context.Crews.Where(c => c.Id == id).SingleOrDefaultAsync();
+            if (crew == null) throw new ArgumentException("crew does not exist!");
+            crew!.FirstName = dto.FirstName.IsNullOrEmpty() ? crew.FirstName : dto.FirstName;
+            crew.LastName = dto.LastName.IsNullOrEmpty() ? crew.LastName : dto.LastName;
+            crew.MiddleName = dto.MiddleName.IsNullOrEmpty() ? crew.MiddleName : dto.MiddleName;
+            crew.BirthDate = dto.BirthDate ?? crew.BirthDate;
+            crew.LastModifiedOn = DateTime.UtcNow;
+            _context.Crews.Update(crew);
+            await _context.SaveChangesAsync();
+        }
     }
 
-    public async Task DeleteCrew(int id)
+    public async Task DeleteCrew(int id,int userId)
     {
-        var crew = await _context.Crews.Where(c => c.Id == id).SingleOrDefaultAsync();
-        _context.Crews.Remove(crew!);
-        await _context.SaveChangesAsync();
+        //only admins can delete Crew so this checks to make sure 
+        var isUserAuthorized = await _userRepository.IsUserAdmin(userId);
+        if (isUserAuthorized)
+        {
+            var crew = await _context.Crews.Where(c => c.Id == id).SingleOrDefaultAsync();
+            _context.Crews.Remove(crew!);
+            await _context.SaveChangesAsync();
+        }
     }
 }
