+++ MovieReviewSite.Core/Repositories/Movie/MovieRepository.cs
@@ -11,11 +11,14 @@ public partial class MovieRepository : IMovieRepository
 {
     private readonly ReviewSiteContext _context;
     private readonly IReviewRepository _reviewRepository;
+    private readonly IUserRepository _userRepository;
 
-    public MovieRepository(ReviewSiteContext context, IReviewRepository reviewRepository)
+    public MovieRepository(ReviewSiteContext context, IReviewRepository reviewRepository,
+        IUserRepository userRepository)
     {
         _context = context;
         _reviewRepository = reviewRepository;
+        _userRepository = userRepository;
     }
 
     public async Task<MovieBase?> GetMovieById(int id)
@@ -26,51 +29,65 @@ public partial class MovieRepository : IMovieRepository
             Name = m.Name
         }).SingleOrDefaultAsync();
     }
-    
-    public  async Task<DataBase.Movie?> GetById(int id)
+
+    public async Task<DataBase.Movie?> GetById(int id)
     {
         return await _context.Movies.Where(m => m.Id == id).SingleOrDefaultAsync();
     }
 
-    public async Task AddMovie(NewMovie movie)
+    public async Task AddMovie(NewMovie dto)
     {
-        var newMovie = new DataBase.Movie()
+        //only admins or vip members can add movies so this checks to make sure 
+        var isUserAuthorized = await _userRepository.IsUserAdminOrVIP(dto.UserId);
+        if (isUserAuthorized)
         {
-            Name = !movie.Name.IsNullOrEmpty() ? movie.Name : null,
-            Synopsis = !movie.Synopsis.IsNullOrEmpty() ? movie.Synopsis : null,
-            Duration = movie.Duration == 0 ? movie.Duration : 0,
-            AgeRateId = movie.AgeRate == 0 ? movie.AgeRate : 0,
-            StatusId = 1,
-            CreatedOn = DateTime.UtcNow,
-            LastModifiedOn = DateTime.UtcNow,
-            RealeaseDate = movie.ReleaseDate == null ? null : movie.ReleaseDate,
-            CreatedBy = movie.CreatedById
-        };
-        await _context.Movies.AddAsync(newMovie);
-        await _context.SaveChangesAsync();
+            var newMovie = new DataBase.Movie()
+            {
+                Name = !dto.Name.IsNullOrEmpty() ? dto.Name : null,
+                Synopsis = !dto.Synopsis.IsNullOrEmpty() ? dto.Synopsis : null,
+                Duration = dto.Duration == 0 ? dto.Duration : 0,
+                AgeRateId = dto.AgeRate == 0 ? dto.AgeRate : 0,
+                StatusId = 1,
+                CreatedOn = DateTime.UtcNow,
+                LastModifiedOn = DateTime.UtcNow,
+                RealeaseDate = dto.ReleaseDate == null ? null : dto.ReleaseDate,
+                CreatedBy = dto.UserId,
+            };
+            await _context.Movies.AddAsync(newMovie);
+            await _context.SaveChangesAsync();
+        }
     }
 
-    public async Task UpdateMovie(int id,UpdatedMovie dto)
+    public async Task UpdateMovie(int id, UpdatedMovie dto)
     {
-        var movie = await _context.Movies.Where(m => m.Id == id).SingleOrDefaultAsync();
+        //only admins or vip members can update movies so this checks to make sure 
+        var isUserAuthorized = await _userRepository.IsUserAdminOrVIP(dto.UserId);
+        if (isUserAuthorized)
+        {
+            var movie = await _context.Movies.Where(m => m.Id == id).SingleOrDefaultAsync();
 
-        if (movie == null) throw new ArgumentException("this movie does not exist!");
-        movie.Name = dto.Name.IsNullOrEmpty() ? movie.Name : dto.Name ;
-        movie.Duration = dto.Duration == 0 ? movie.Duration : dto.Duration;
-        movie.RealeaseDate = dto.ReleaseDate ?? movie.RealeaseDate;
-        movie.LastModifiedOn = DateTime.UtcNow;
-        movie.AgeRateId = dto.AgeRate == 0 ? movie.AgeRateId : dto.AgeRate;
-        movie.Synopsis = dto.Synopsis.IsNullOrEmpty() ? movie.Synopsis : dto.Synopsis;
+            if (movie == null) throw new ArgumentException("this movie does not exist!");
+            movie.Name = dto.Name.IsNullOrEmpty() ? movie.Name : dto.Name;
+            movie.Duration = dto.Duration == 0 ? movie.Duration : dto.Duration;
+            movie.RealeaseDate = dto.ReleaseDate ?? movie.RealeaseDate;
+            movie.LastModifiedOn = DateTime.UtcNow;
+            movie.AgeRateId = dto.AgeRate == 0 ? movie.AgeRateId : dto.AgeRate;
+            movie.Synopsis = dto.Synopsis.IsNullOrEmpty() ? movie.Synopsis : dto.Synopsis;
 
-        _context.Update(movie);
-        await _context.SaveChangesAsync();
-
+            _context.Update(movie);
+            await _context.SaveChangesAsync();
+        }
     }
 
-    public async Task DeleteMovie(int id)
+    public async Task DeleteMovie(int id,int userId)
     {
-        var movie = await _context.Movies.Where(m => m.Id == id).SingleOrDefaultAsync();
-        _context.Movies.Remove(movie!);
-        await _context.SaveChangesAsync();
+        //only admins can delete movies so this checks to make sure 
+        var isUserAuthorized = await _userRepository.IsUserAdmin(userId);
+        if (isUserAuthorized)
+        {
+            var movie = await _context.Movies.Where(m => m.Id == id).SingleOrDefaultAsync();
+            _context.Movies.Remove(movie!);
+            await _context.SaveChangesAsync();
+        }
     }
 }
